parallel-processing/requirements.txt
click==8.1.3
google-cloud-storage==2.3.0